axizさんが変更

+++ src/main/java/com/example/shimakiti/controller/CategoryController.java
@@ -1,19 +1,12 @@
 package com.example.shimakiti.controller;
 
 import com.example.shimakiti.service.CategoryService;
-import com.example.shimakiti.service.ShimaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import com.example.shimakiti.entity.Categories;
 import com.example.shimakiti.entity.CategoryDetail;
-import com.example.shimakiti.service.CategoryService;
-import com.example.shimakiti.service.ShimaService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -30,9 +23,9 @@ public class CategoryController {
         var list = new ArrayList<CategoryDetail>();
         for(Categories categories:categoryService.findCategory()){
             CategoryDetail categoryDetail = new CategoryDetail();
-            categoryDetail.setId((long) categories.getId());
+            categoryDetail.setId(categories.getId());
             categoryDetail.setName(categories.getName());
-            categoryDetail.setPosts((long) categoryService.findByCategoryPosts(categories).size());
+            categoryDetail.setPosts(categoryService.findByCategoryPosts(categories).size());
             list.add(categoryDetail);
         }
         model.addAttribute("categories",list);
@@ -44,7 +37,7 @@ public class CategoryController {
         var categories = categoryService.findById(id);
 
         CategoryDetail categoryDetail = new CategoryDetail();
-        categoryDetail.setId((long) categories.getId());
+        categoryDetail.setId(categories.getId());
         categoryDetail.setName(categories.getName());
         categoryDetail.setPosts((long) categoryService.findByCategoryPosts(categories).size());
